requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
chromedriver-autoinstaller==0.4.0
docx==0.2.4
exceptiongroup==1.1.0
Faker==16.6.1
fpdf==1.7.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
outcome==1.2.0
ow==0.0.4
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pypdf2==3.0.1
PySocks==1.7.1
pytest==7.2.1
pytest-dependency==0.5.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
pytest-xdist==2.5.0
